requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
django-cors-headers==4.3.1
whitenoise==6.6.0

# PostgreSQL adapter (install manually if using PostgreSQL)
# psycopg2-binary==2.9.9  # Uncomment and install if using PostgreSQL
